Enum Profession to Entity

+++ Project/src/main/java/de/wwweasel/RCP/EmployeeController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import DTO.EmployeeDTO;
+import DTO.EmployeesDTO;
 import DTO.EmployeeProfesioNDTO;
 
 @Controller
@@ -20,19 +20,26 @@ public class EmployeeController {
 	ProfessioNService professioNService;
 
 	@RequestMapping(method=RequestMethod.GET,value="/")
-	public String start( Model model, @RequestParam(required = false, defaultValue = "none") String filter, @RequestParam(required = false) Integer[] employeeIds) {//, @ModelAttribute("editError") String editError
+	public String start( Model model, @RequestParam(required = false, defaultValue = "removeFilter") String filter, @RequestParam(required = false) Integer[] employeeIds) {//, @ModelAttribute("editError") String editError
 		
 		switch (filter) {
 		case "findByProfession":
-			//model.addAttribute("employees", employeeService.findByProfession(employeeIds));
-			model.addAttribute("employees", employeeService.findAll());
+			if(employeeIds!=null) {
+				model.addAttribute("employees", employeeService.findByProfession(employeeIds));
+			}else {
+				model.addAttribute("employees", employeeService.findAll());
+			}
 			break;
 			
 		case "findByIds":
-			model.addAttribute("employees", employeeService.findByIds(employeeIds).getEmployees());
+			if(employeeIds!=null) {
+				model.addAttribute("employees", employeeService.findByIds(employeeIds).getEmployees());
+			}else {
+				model.addAttribute("employees", employeeService.findAll());
+			}
 			break;
 			
-		case "none":
+		case "removeFilter":
 			model.addAttribute("employees", employeeService.findAll());
 			break;
 			
@@ -46,34 +53,24 @@ public class EmployeeController {
 	}
 	
 	@RequestMapping(method=RequestMethod.GET,value="/add")
-	public String addEmployee(Model model) {
-		EmployeeProfesioNDTO employeeProfesioNDTO = new EmployeeProfesioNDTO();
-		employeeProfesioNDTO.setEmployee(new Employee());
-		employeeProfesioNDTO.setProfessioN(new ProfessioN());
-		
-		model.addAttribute("employeeProfesioNDTO", employeeProfesioNDTO);
-		model.addAttribute("professions", Professions.values() ); // Feed all Enum COnstants for the Dropdown
+	public String addEmployee(Model model) {		
+		model.addAttribute("employee", new Employee());
+		model.addAttribute("professions", professioNService.findAll() );
 		return "add";
 	}
 	
 	@RequestMapping(method=RequestMethod.POST,value="/add")
-	public String addEmployee(@ModelAttribute EmployeeProfesioNDTO employeeProfesioNDTO) {
-		ProfessioN professioN = employeeProfesioNDTO.getProfessioN();
-		Employee employee = employeeProfesioNDTO.getEmployee();
-		
-		professioN = professioNService.save(professioN);
-		
-		employee.setProfession(professioN);
+	public String addEmployee(@ModelAttribute Employee employee) {
 		employeeService.save( employee );
 		return "redirect:/";
 	}
 	
 	@RequestMapping(method=RequestMethod.GET,value="/edit")
 	public String editEmployee(@RequestParam(required = false) Integer[] employeeIds, Model model) {
-		
 		if(employeeIds!=null) {
-			EmployeeDTO dto = employeeService.findByIds(employeeIds);
+			EmployeesDTO dto = employeeService.findByIds(employeeIds);
 			model.addAttribute("employeesDTO",dto);
+			model.addAttribute("professions", professioNService.findAll() );
 			return "edit";
 		}else {
 			return "redirect:/";
@@ -81,7 +78,7 @@ public class EmployeeController {
 	}
 	
 	@RequestMapping(method=RequestMethod.POST,value="/edit")
-	public String editEmployee(@ModelAttribute EmployeeDTO employeeDTO) {
+	public String editEmployee(@ModelAttribute EmployeesDTO employeeDTO) {
 		for (Employee employee : employeeDTO.getEmployees()) {
 			System.out.println("Save: " + employee);
 			employeeService.save(employee);
@@ -91,9 +88,11 @@ public class EmployeeController {
 	
 	
 	@RequestMapping(method=RequestMethod.GET,value="/delete")
-	public String deleteEmployee(@RequestParam Integer[] employeeIds) {
-		for (int i = 0; i < employeeIds.length; i++) {
-			employeeService.deleteById(employeeIds[i]);
+	public String deleteEmployee(@RequestParam(required = false) Integer[] employeeIds) {
+		if(employeeIds!=null) {
+			for (int i = 0; i < employeeIds.length; i++) {
+				employeeService.deleteById(employeeIds[i]);
+			}
 		}
 		return "redirect:/";
 	}
